+++ variables.tf
@@ -3,12 +3,8 @@ variable "ip_blacklist" {
   type    = list(string)
 
   description = <<-END
-    The IP Blacklist 
-
-    The IP Blacklist defines a set of IP addresses that will be trusted.
-
-    For cases where a given IP is in both the IP Whitelist and the IP Blacklist
-    requests will be blocked because the IP Blacklist takes precidence.
+    a list of IP addresses that will be blocked.
+    https://go.s3d.club/tf/waf#ip_blacklist
     END
 }
 
@@ -17,12 +13,17 @@ variable "ip_whitelist" {
   type    = list(string)
 
   description = <<-END
-    The IP Whitelist
+    a list of IP addresses that will be allowed.
+    https://go.s3d.club/tf/waf#ip_whitelist
+    END
+}
 
-    The IP Whitelist defines a set of IP addresses that will be trusted.
+variable "kms_key_arn" {
+  type = string
 
-    For cases where a given IP is in both the IP Whitelist and the IP Blacklist
-    requests will be blocked because the IP Blacklist takes precidence.
+  description = <<-END
+    KMS Key ARN
+    https://go.s3d.club/tf/waf#kms_key_arn
     END
 }
 
@@ -30,11 +31,8 @@ variable "name_prefix" {
   type = string
 
   description = <<-END
-    The name of the WAF resource
-
-    The WAF resource defined by this module will have this name. The names of
-    associated resources incorporate the name of the WAF resource as a prefix in
-    cases where they need unique names.
+    a prefix for resource names.
+    https://go.s3d.club/tf/waf#name_prefix
     END
 }
 
@@ -43,8 +41,8 @@ variable "redirects" {
   type    = map(string)
 
   description = <<-END
-		Redirects
-		https://go.s3d.club/waf#redirects
+    a map of path with urls for 301 redirects.
+    https://go.s3d.club/waf#redirects
     END
 }
 
@@ -53,8 +51,7 @@ variable "tags" {
   type    = map(string)
 
   description = <<-END
-    The tags for resources
-
-    This module will include these tags on all resources.
+    A map of tags for resources.
+    https://go.s3d.club/waf#redirects
     END
 }